+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 2.0.0 - 2023-10-26
+## 2.1.0 - 2024-03-10
+
+### Added
+
+- Support for `innmind/operating-system:~5.0`
+- Support for `innmind/async-http-server:~3.0`
+
+## 2.0.0 - 2023-11-26
 
 ### Added